Use updated types in metrics app
