--- packages/metrics-app/src/types/query.ts
@@ -1,10 +1,0 @@
-export interface Query {
-  id: string
-  measurements: QueryMeasurementType[]
-}
-
-export enum QueryMeasurementType {
-  Bandwidth = 'bandwidth',
-  Latency = 'latency',
-  SignalStrength = 'signalStrength'
-}
